Move station select map into  it's own reuseable component and move the various hydstra pieces to their own types to make them reusable as well
