+++ Source/Nebula.Web/src/app/shared/components/station-select-map/station-select-map.component.html
@@ -1,0 +1,2 @@
+<div #mapDiv id={{mapID}} class="location-card" [style.min-height]="mapHeight"
+                                style="height:100%"></div>
